Avoid static values in field maps
