+++ src/result_processor.c
@@ -605,8 +605,7 @@ int loader_Next(ResultProcessorCtx *ctx, SearchResult *r) {
   // TODO: load should return strings, not redis strings
   for (int i = 0; i < doc.numFields; i++) {
     if (doc.fields[i].text) {
-      RSFieldMap_SetRawValue(&r->fields, doc.fields[i].name, RSValue_RedisString,
-                             doc.fields[i].text);
+      RSFieldMap_Set(&r->fields, doc.fields[i].name, RS_RedisStringVal(doc.fields[i].text));
     } else {
       RSFieldMap_Set(&r->fields, doc.fields[i].name, RS_NullVal());
     }
